requirements.txt
fastapi~=0.115.5
numpy~=2.2.0
pandas~=2.2.3
torch~=2.5.1
environs~=11.2.1
python-dateutil~=2.9.0.post0
transformers~=4.47.0
scikit-learn~=1.6.0